requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
fpdf==1.7.2
idna==3.10
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
psutil==6.1.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
requests-futures==1.0.1
setuptools==75.2.0
sherlock-project==0.15.0
six==1.16.0
stem==1.8.2
tzdata==2024.2
urllib3==2.2.3
